Un-feature flag new asset details
