+++ js_modules/dagit/packages/core/src/app/Flags.tsx
@@ -9,7 +9,6 @@ export const DAGIT_FLAGS_KEY = 'DAGIT_FLAGS';
 export const FeatureFlag = {
   flagDebugConsoleLogging: 'flagDebugConsoleLogging' as const,
   flagDisableWebsockets: 'flagDisableWebsockets' as const,
-  flagNewAssetDetails: 'flagNewAssetDetails' as const,
   flagAssetGraphExperimentalZoom: 'flagAssetGraphExperimentalZoom' as const,
 };
 export type FeatureFlagType = keyof typeof FeatureFlag;
