+++ js_modules/dagit/packages/core/src/app/UserSettingsRoot.tsx
@@ -126,16 +126,6 @@ export function getFeatureFlagRows(
       ),
     },
     {
-      key: 'New asset detail pages',
-      value: (
-        <Checkbox
-          format="switch"
-          checked={flags.includes(FeatureFlag.flagNewAssetDetails)}
-          onChange={() => toggleFlag(FeatureFlag.flagNewAssetDetails)}
-        />
-      ),
-    },
-    {
       key: 'Experimental "groups-only" asset graph zoom level',
       value: (
         <Checkbox
