+++ js_modules/dagit/packages/core/src/assets/AssetView.test.tsx
@@ -10,10 +10,15 @@ import {AssetView} from './AssetView';
 // This file must be mocked because Jest can't handle `import.meta.url`.
 jest.mock('../graph/asyncGraphLayout', () => ({}));
 
+// This file must be mocked because useVirtualizer tries to create a ResizeObserver,
+// and the component tree fails to mount.
+jest.mock('./AssetPartitions', () => ({AssetPartitions: () => <div />}));
+
 describe('AssetView', () => {
   const mocks = {
     AssetNode: () => ({
       partitionKeys: () => [...new Array(0)],
+      partitionKeysByDimension: () => [...new Array(1)],
     }),
     Asset: () => ({
       assetMaterializations: () => [...new Array(1)],
