+++ js_modules/dagit/packages/core/src/assets/AssetView.tsx
@@ -14,7 +14,6 @@ import {
 import * as React from 'react';
 import {Link} from 'react-router-dom';
 
-import {useFeatureFlags} from '../app/Flags';
 import {
   FIFTEEN_SECONDS,
   QueryRefreshCountdown,
@@ -33,7 +32,6 @@ import {buildRepoAddress} from '../workspace/buildRepoAddress';
 import {workspacePathFromAddress} from '../workspace/workspacePath';
 
 import {AssetEvents} from './AssetEvents';
-import {AssetEventsV2} from './AssetEventsV2';
 import {AssetNodeDefinition, ASSET_NODE_DEFINITION_FRAGMENT} from './AssetNodeDefinition';
 import {AssetNodeInstigatorTag, ASSET_NODE_INSTIGATORS_FRAGMENT} from './AssetNodeInstigatorTag';
 import {AssetNodeLineage} from './AssetNodeLineage';
@@ -70,9 +68,7 @@ export const AssetView: React.FC<Props> = ({assetKey}) => {
     assetKey,
   );
 
-  const {flagNewAssetDetails} = useFeatureFlags();
-  const defaultTab =
-    flagNewAssetDetails && definition?.partitionDefinition ? 'partitions' : 'events';
+  const defaultTab = definition?.partitionDefinition ? 'partitions' : 'events';
   const selectedTab = params.view || defaultTab;
 
   // Load the asset graph - a large graph for the Lineage tab, a small graph for the Definition tab
@@ -157,33 +153,19 @@ export const AssetView: React.FC<Props> = ({assetKey}) => {
         tabs={
           <Box flex={{direction: 'row', justifyContent: 'space-between', alignItems: 'flex-end'}}>
             <Tabs size="large" selectedTabId={selectedTab}>
-              {flagNewAssetDetails && definition?.partitionDefinition && (
+              {definition?.partitionDefinition && (
                 <Tab
                   id="partitions"
                   title="Partitions"
                   onClick={() => setParams({...params, view: 'partitions'})}
                 />
               )}
-              {flagNewAssetDetails ? (
-                <Tab
-                  id="events"
-                  title="Events"
-                  onClick={() => setParams({...params, view: 'events'})}
-                />
-              ) : (
-                <Tab
-                  id="events"
-                  title="Activity"
-                  onClick={() => setParams({...params, view: 'events'})}
-                />
-              )}
-              {flagNewAssetDetails && (
-                <Tab
-                  id="plots"
-                  title="Plots"
-                  onClick={() => setParams({...params, view: 'plots'})}
-                />
-              )}
+              <Tab
+                id="events"
+                title="Events"
+                onClick={() => setParams({...params, view: 'events'})}
+              />
+              <Tab id="plots" title="Plots" onClick={() => setParams({...params, view: 'plots'})} />
               <Tab
                 id="definition"
                 title="Definition"
@@ -212,6 +194,14 @@ export const AssetView: React.FC<Props> = ({assetKey}) => {
           </Box>
         }
       />
+      {!viewingMostRecent && (
+        <HistoricalViewAlert
+          asOf={params.asOf}
+          onClick={() => setParams({asOf: undefined, time: params.asOf})}
+          hasDefinition={!!definition}
+        />
+      )}
+
       {
         // Avoid thrashing the events UI (which chooses a different default query based on whether
         // data is partitioned) by waiting for the definition to be loaded before we show any tab content
@@ -225,19 +215,11 @@ export const AssetView: React.FC<Props> = ({assetKey}) => {
         </Box>
       ) : (
         <>
-          {!viewingMostRecent && (
-            <HistoricalViewAlert
-              asOf={params.asOf}
-              onClick={() => setParams({asOf: undefined, time: params.asOf})}
-              hasDefinition={!!definition}
-            />
-          )}
-
           {selectedTab === 'definition' ? (
             renderDefinitionTab()
           ) : selectedTab === 'lineage' ? (
             renderLineageTab()
-          ) : selectedTab === 'partitions' && flagNewAssetDetails ? (
+          ) : selectedTab === 'partitions' ? (
             <AssetPartitions
               assetKey={assetKey}
               assetPartitionNames={definition?.partitionKeysByDimension.map((k) => k.name)}
@@ -247,20 +229,11 @@ export const AssetView: React.FC<Props> = ({assetKey}) => {
               setParams={setParams}
               liveData={definition ? liveDataByNode[toGraphId(definition.assetKey)] : undefined}
             />
-          ) : selectedTab === 'events' && flagNewAssetDetails ? (
-            <AssetEventsV2
-              assetKey={assetKey}
-              assetLastMaterializedAt={lastMaterializedAt}
-              params={params}
-              paramsTimeWindowOnly={!!params.asOf}
-              setParams={setParams}
-              liveData={definition ? liveDataByNode[toGraphId(definition.assetKey)] : undefined}
-            />
           ) : selectedTab === 'events' ? (
             <AssetEvents
               assetKey={assetKey}
-              assetLastMaterializedAt={lastMaterializedAt}
               assetHasDefinedPartitions={!!definition?.partitionDefinition}
+              assetLastMaterializedAt={lastMaterializedAt}
               params={params}
               paramsTimeWindowOnly={!!params.asOf}
               setParams={setParams}
